% Generated by roxygen2 (4.0.1): do not edit by hand
\name{json_to_list}
\alias{json_to_list}
\title{Convert JSON string to a list().}
\usage{
json_to_list(json)
}
\arguments{
\item{json}{String representing a JSON object}
}
\value{
A list() representing the JSON object
}
\description{
Convert JSON string to a list().
}
\examples{
L <- json_to_list("{a:1,b:["x","y","z"]}")
}

