% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_bson}
\alias{make_bson}
\title{Adds new item to mongo.bson.buffer}
\usage{
make_bson(obj, name, value, type)
}
\arguments{
\item{obj}{Initial mongo.bson.buffer}

\item{name}{String holding name of node}

\item{value}{logical, numeric or string holding  value of node}

\item{type}{Identifier for type of node}
}
\value{
mongo.bson.buffer with new item added
}
\description{
Adds new item to mongo.bson.buffer
}

