% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_df}
\alias{make_df}
\title{Adds new line to a data.frame()}
\usage{
make_df(obj, name, value, type)
}
\arguments{
\item{obj}{Initial data.frame()}

\item{name}{String holding name of node}

\item{value}{logical, numeric or string holding  value of node}

\item{type}{Identifier for type of node}
}
\value{
obj with line added representing new node
}
\description{
Adds new line to a data.frame()
}

