% Generated by roxygen2 (4.0.1): do not edit by hand
\name{traverse_list_recursively}
\alias{traverse_list_recursively}
\title{Traverse list recursively and apply function to nodes.}
\usage{
traverse_list_recursively(l, obj, fun)
}
\arguments{
\item{l}{list() to be traversed}

\item{obj}{Initial object which is supposed to be returned}

\item{fun}{Function to be applied to nodes}
}
\value{
An object of same type as obj made by fun
}
\description{
Traverse list recursively and apply function to nodes.
}
\examples{
l <- list(a=1,b=list(c=2,d=3))
L <- json_to_list(l = l, obj = data.frame(), fun = make_df)
}

